Rename ./src/steps/scans -> ./src/steps/vulnerabilities
